Tidy constanst and factories
