+++ pipelex/plugins/portkey/portkey_factory.py
@@ -10,7 +10,7 @@ from pipelex.hub import get_telemetry_manager
 from pipelex.plugins.openai.openai_constants import OpenAIBodyKey
 from pipelex.plugins.portkey.portkey_constants import PortkeyHeaderKey
 from pipelex.plugins.portkey.portkey_exceptions import PortkeyCredentialsError
-from pipelex.system.telemetry.otel_constants import PIPE_CODE_REDACTED
+from pipelex.system.telemetry.otel_constants import OTelConstants
 from pipelex.system.telemetry.telemetry_manager_abstract import TelemetryManagerAbstract
 
 if TYPE_CHECKING:
@@ -56,7 +56,7 @@ class PortkeyFactory:
             # Build span name respecting privacy settings
             pipe_code = llm_job.job_metadata.pipe_code or "main"
             if not TelemetryManagerAbstract.is_capture_pipe_codes_enabled():
-                pipe_code = PIPE_CODE_REDACTED
+                pipe_code = OTelConstants.PIPE_CODE_REDACTED
 
             unit_job_id = llm_job.job_metadata.unit_job_id or "unknown"
             extra_headers[PortkeyHeaderKey.SPAN_NAME] = f"{pipe_code}: {unit_job_id} -> {output_desc}"
